Update SNOMEDCT Version Source
The previous source for SNOMEDCT, as specified in
https://documentation.uts.nlm.nih.gov/automating-downloads.html, was
outdated and pinned to an older version.

To ensure up-to-date data, we have switched
to using https://www.nlm.nih.gov/healthit/snomedct/us_edition.html as
our new source, we've implemented a scraper that extracts the download
link from the "Download Now!" button on the webpage.

Co-authored-by: @Panthevm <panthevm@yandex.ru>